map rendering on home page
