--- client/src/app.css
@@ -1,4 +1,0 @@
-.leaflet-container {
-    width: 100vh;
-    height: 100vh;
-}
