+++ client/src/components/Mapbox/index.css
@@ -1,0 +1,6 @@
+.leaflet-container {
+  height: 50vh;
+  padding: 5px;
+  border: 1px solid #ccc;
+  margin: 2px;
+}
