+++ client/src/pages/Home.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useState } from "react";
 import Nav from "../components/Nav";
+import Mapbox from "../components/Mapbox";
 
 
 const Home = () => {
@@ -8,6 +9,7 @@ const Home = () => {
   return (
     <div>
       <h1>Urban Oasis</h1>
+      <Mapbox />
     </div>
     
   );
